Made GLFW MTd and MT, tidied up how those flags are set in the CMakeLists.txt in root. Window has a public "Closed" method now. Fixed a small crash in V8 on shutdown. Added 'quit' as console command. Moved logs from snuffbox-test app to actual SnuffboxApp

+++ snuffbox-engine/js/js_wrapper.h
@@ -292,21 +292,21 @@ namespace snuffbox
 		template<>
         inline v8::Local<v8::Value> JSWrapper::CastValue<int>(const int& val)
 		{
-			return CastValue<int>(static_cast<double>(val));
+			return CastValue<double>(static_cast<double>(val));
 		}
 
 		//-------------------------------------------------------------------------------------------
 		template<>
 		inline v8::Local<v8::Value> JSWrapper::CastValue<unsigned int>(const unsigned int& val)
 		{
-			return CastValue<unsigned int>(static_cast<double>(val));
+			return CastValue<double>(static_cast<double>(val));
 		}
 
 		//-------------------------------------------------------------------------------------------
 		template<>
         inline v8::Local<v8::Value> JSWrapper::CastValue<float>(const float& val)
 		{
-			return CastValue<float>(static_cast<double>(val));
+			return CastValue<double>(static_cast<double>(val));
 		}
 
 		//-------------------------------------------------------------------------------------------
